added createdBy field

+++ src/Graph/resolvers/player.ts
@@ -9,13 +9,13 @@ import { mongoUser } from '../../MongoDB/User'
 import { checkPrivileges } from '../../Middleware/isAuth'
 import { Player, PlayerType } from '../../MongoDB/Player/Entities'
 import moment from 'moment'
-import { gql_User, gql_Player, playerLoader } from './transform'
+import { gql_Player, playerLoader } from './transform'
 
 const playerResolver = {
   Query: {
     getPlayers: async (_, { playerFilters }, { req }) => {
       if (!req.isAuth) throw new Error(ErrorMessages.user_unauthenticated)
-      const { result, ...rest }: List<Player> = await mongoPlayer.getPlayers(playerFilters)
+      const { result, ...rest }: List<Player> = await mongoPlayer.getPlayers(playerFilters, req.idUser)
       return {
         ...rest,
         result: result.map(gql_Player)
@@ -31,9 +31,9 @@ const playerResolver = {
         if (!userId && (!userData || isEmpty(userData))) throw new Error(ErrorMessages.player_user_not_specified)
         let idUser = userId
         if (!idUser) {
-          idUser = await mongoUser.createUser({ ...userData, privilege: Privilege.User })
+          idUser = await mongoUser.createUser({ ...userData, privilege: Privilege.User }, req.idUser)
         }
-        const idPlayer = await mongoPlayer.createPlayer({ ...playerData, idUser })
+        const idPlayer = await mongoPlayer.createPlayer({ ...playerData, idUser }, req.idUser)
         return idPlayer
       } catch (error) {
         throw error
@@ -41,6 +41,7 @@ const playerResolver = {
     },
     updatePlayer: async (_, { updatePlayerInput }, { req }) => {
       if (!req.isAuth) throw new Error(ErrorMessages.user_unauthenticated)
+      checkPrivileges(req)
       const { _id, positions, state, score } = updatePlayerInput
 
       if (isEmpty(cleanDeep(updatePlayerInput))) return true
@@ -52,7 +53,7 @@ const playerResolver = {
       updatedPlayer.updatedAt = moment().toDate()
 
       const { modifiedCount } = await MongoDBInstance.collection.player.updateOne(
-        { _id: new ObjectId(_id) },
+        { _id: new ObjectId(_id), createdBy: new ObjectId(req.idUser) },
         { $set: updatedPlayer }
       )
       
@@ -62,14 +63,15 @@ const playerResolver = {
     },
     deletePlayer: async (_, { deletePlayerInput }, { req }) => {
       if (!req.isAuth) throw new Error(ErrorMessages.user_unauthenticated)
+      checkPrivileges(req)
       const { _id, idUser, type } = deletePlayerInput
       const promises = []
 
       // delete player from player collection
-      promises.push(MongoDBInstance.collection.player.deleteOne({ _id: new ObjectId(_id), user: new ObjectId(idUser) }))
+      promises.push(MongoDBInstance.collection.player.deleteOne({ _id: new ObjectId(_id), user: new ObjectId(idUser), createdBy: new ObjectId(req.idUser) }))
       // delete player from user collection
       promises.push(MongoDBInstance.collection.user.updateOne(
-        { _id: new ObjectId(idUser) },
+        { _id: new ObjectId(idUser), createdBy: new ObjectId(req.idUser) },
         { $set: { ...type === PlayerType.Football
           ? { footballPlayer: null }
           : { futsalPlayer: null }
